MAG-378: Refactor OAuth2 service and admin configuration handling
